feat(EST-3151-FE): validate tool parameters JSON against schema
Validate the custom tool parameters JSON against a JSON Schema via Monaco,
surfacing errors inline with line and column. Track the last valid JSON on
every change and add a revert action to restore it. Guard the table-mode
switch on schema validity so invalid JSON no longer silently empties the
table. Reorganize the parameters config into a parameters/ folder (model,
serialization, validation, columns).
