+++ frontend/src/app/user-settings-page/tools/custom-tool-editor/create-custom-tool-dialog/create-custom-tool-dialog.component.html
@@ -148,6 +148,11 @@
                             <app-help-tooltip
                                 text="Configure the JSON parameters your custom tool accepts."
                             ></app-help-tooltip>
+                            @if (!parametersTableMode() && jsonIssues().length) {
+                                <span class="parameters-issue-label">
+                                    Issues detected. Switching modes may lose changes
+                                </span>
+                            }
                         </div>
                         <div class="parameters-mode-switch">
                             <span>Table mode</span>
@@ -180,6 +185,20 @@
                                         [text]="variablesSchemaTooltip"
                                         position="top"
                                     ></app-help-tooltip>
+                                    @if (jsonIssues().length && lastValidJson()) {
+                                        <button
+                                            type="button"
+                                            class="revert-action"
+                                            title="Revert to the last valid JSON — the most recent version with no errors"
+                                            aria-label="Revert to the last valid JSON"
+                                            (click)="revertToLastValidJson()"
+                                        >
+                                            <app-svg-icon
+                                                icon="refresh"
+                                                size="1rem"
+                                            ></app-svg-icon>
+                                        </button>
+                                    }
                                 </div>
                                 <div class="editor-card-actions">
                                     <button
@@ -221,9 +240,11 @@
                                     <app-json-editor
                                         [showHeader]="false"
                                         [fullHeight]="true"
+                                        [jsonSchema]="toolVariablesSchema"
                                         [jsonData]="form.controls.variablesJson.value"
                                         (jsonChange)="onJsonChange($event)"
                                         (validationChange)="onJsonValidationChange($event)"
+                                        (errorsChange)="onJsonErrorsChange($event)"
                                         (editorReady)="onJsonEditorReady($event)"
                                     ></app-json-editor>
                                 </div>
@@ -311,8 +332,10 @@
                                 [showHeader]="false"
                                 [jsonData]="form.controls.variablesJson.value"
                                 [fullHeight]="true"
+                                [jsonSchema]="toolVariablesSchema"
                                 (jsonChange)="onJsonChange($event)"
                                 (validationChange)="onJsonValidationChange($event)"
+                                (errorsChange)="onJsonErrorsChange($event)"
                                 (editorReady)="onJsonEditorReady($event)"
                             ></app-json-editor>
                         }
