+++ frontend/src/index.html
@@ -502,6 +502,24 @@
             </symbol>
 
             <symbol
+                id="icon-refresh"
+                viewBox="0 0 24 24"
+                fill="none"
+                stroke="currentColor"
+                stroke-width="2"
+                stroke-linecap="round"
+                stroke-linejoin="round"
+            >
+                <path
+                    stroke="none"
+                    d="M0 0h24v24H0z"
+                    fill="none"
+                />
+                <path d="M20 11a8.1 8.1 0 0 0 -15.5 -2m-.5 -4v4h4" />
+                <path d="M4 13a8.1 8.1 0 0 0 15.5 2m.5 4v-4h-4" />
+            </symbol>
+
+            <symbol
                 id="icon-database"
                 viewBox="0 0 24 24"
                 fill="none"
